Properly recognize html element as root to fix layout bug
Before this change, recognizing the 'document' element as root and 'html' as non-root resulted in all elements in the document having a 'start_y' of viewport height, which isn't right and resulted in nothing being visible besides the background.  This commit changes the code to properly recognize html as the root of the page.